fix(pid_utils): match pidfile from the provided pid_path
kill_and_wait function expects to receive absolute path to pidfile
while currently asterisk character used to create the pattern.

This commit adds additional check to kill_and_wait function that will
find an absolute path to the process pidfile. If the pidfile is not
found, which means that process is not running, then the function will
return exit code 0.